[onert] Move shape calculation code into ShapeInference.cc
For Op named "ABC", this commit moves inferABCShape(..) into ShapeInference.cc.

After this work, static shape inferer code will be moved to compiler dir,
and dynamic shape inferer code will be moved to exec dir.
