+++ runtime/onert/core/src/util/shapeinf/Select.cc
@@ -21,85 +21,6 @@ namespace onert
 namespace shape_inference
 {
 
-bool HaveSameShapes(const ir::Shape &input_cond_shape, const ir::Shape &input_true_shape,
-                    const ir::Shape &input_false_shape)
-{
-  if ((input_cond_shape.rank() != input_true_shape.rank()) ||
-      input_cond_shape.rank() != input_false_shape.rank())
-  {
-    return false;
-  }
-
-  int rank = input_cond_shape.rank();
-  for (int i = 0; i < rank; ++i)
-  {
-    if (input_cond_shape.dim(i) != input_true_shape.dim(i) ||
-        input_cond_shape.dim(i) != input_false_shape.dim(i))
-    {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-bool calculateShape(const ir::Shape &input_cond_shape, const ir::Shape &input_true_shape,
-                    const ir::Shape &input_false_shape, ir::Shape &new_shape)
-{
-  ir::Shape cond_shape = input_cond_shape;
-  ir::Shape true_shape = input_true_shape;
-  ir::Shape false_shape = input_false_shape;
-  int most_rank =
-      (cond_shape.rank() >= true_shape.rank()) && (cond_shape.rank() >= false_shape.rank())
-          ? cond_shape.rank()
-          : (false_shape.rank() >= true_shape.rank() ? false_shape.rank() : true_shape.rank());
-
-  ir::Shape calculate_shape(most_rank);
-
-  cond_shape.extendRank(most_rank);
-  true_shape.extendRank(most_rank);
-  false_shape.extendRank(most_rank);
-
-  for (int i = 0; i < most_rank; ++i)
-  {
-    calculate_shape.dim(i) =
-        (cond_shape.dim(i) >= true_shape.dim(i)) && (cond_shape.dim(i) >= false_shape.dim(i))
-            ? cond_shape.dim(i)
-            : (false_shape.dim(i) >= true_shape.dim(i) ? false_shape.dim(i) : true_shape.dim(i));
-
-    if ((cond_shape.dim(i) != calculate_shape.dim(i) && cond_shape.dim(i) != 1) ||
-        (true_shape.dim(i) != calculate_shape.dim(i) && true_shape.dim(i) != 1) ||
-        (false_shape.dim(i) != calculate_shape.dim(i) && false_shape.dim(i) != 1))
-    {
-      return false;
-    }
-  }
-
-  new_shape = calculate_shape;
-
-  return true;
-}
-
-ir::Shape inferSelectShape(const ir::Shape &input_cond_shape, const ir::Shape &input_true_shape,
-                           const ir::Shape &input_false_shape)
-{
-  bool havesame = HaveSameShapes(input_cond_shape, input_true_shape, input_false_shape);
-  if (havesame)
-  {
-    return input_cond_shape;
-  }
-
-  ir::Shape new_shape;
-  bool possible = calculateShape(input_cond_shape, input_true_shape, input_false_shape, new_shape);
-
-  if (!possible)
-  {
-    throw std::runtime_error("Broadcasting is not possible.");
-  }
-
-  return new_shape;
-}
-
 void StaticInferer::visit(const ir::operation::Select &op)
 {
   const auto input_cond_idx{op.getInputs().at(ir::operation::Select::Input::CONDITION)};
