+++ runtime/onert/core/src/util/shapeinf/StridedSlice.cc
@@ -23,198 +23,6 @@ namespace onert
 {
 namespace shape_inference
 {
-struct StridedSliceParams
-{
-  int8_t start_indices_count;
-  int16_t start_indices[4];
-  int8_t stop_indices_count;
-  int16_t stop_indices[4];
-  int8_t strides_count;
-  int16_t strides[4];
-
-  int16_t begin_mask;
-  int16_t ellipsis_mask;
-  int16_t end_mask;
-  int16_t new_axis_mask;
-  int16_t shrink_axis_mask;
-};
-
-int Clamp(const int v, const int lo, const int hi)
-{
-  assert(!(hi < lo));
-  if (hi < v)
-    return hi;
-  if (v < lo)
-    return lo;
-  return v;
-}
-
-int StartForAxis(const StridedSliceParams &params, const ir::Shape &input_shape, int axis)
-{
-  const auto begin_mask = params.begin_mask;
-  const auto *start_indices = params.start_indices;
-  const auto *strides = params.strides;
-  // Begin with the specified index.
-  int start = start_indices[axis];
-
-  // begin_mask override
-  if (begin_mask & 1 << axis)
-  {
-    if (strides[axis] > 0)
-    {
-      // Forward iteration - use the first element. These values will get
-      // clamped below (Note: We could have set them to 0 and axis_size-1, but
-      // use lowest() and max() to maintain symmetry with StopForAxis())
-      start = std::numeric_limits<int>::lowest();
-    }
-    else
-    {
-      // Backward iteration - use the last element.
-      start = std::numeric_limits<int>::max();
-    }
-  }
-
-  // Handle negative indices
-  int axis_size = input_shape.dim(axis);
-  if (start < 0)
-  {
-    start += axis_size;
-  }
-
-  // Clamping
-  start = Clamp(start, 0, axis_size - 1);
-
-  return start;
-}
-
-// Return the "real" index for the end of iteration along that axis. This is an
-// "end" in the traditional C sense, in that it points to one past the last
-// element. ie. So if you were iterating through all elements of a 1D array of
-// size 4, this function would return 4 as the stop, because it is one past the
-// "real" indices of 0, 1, 2 & 3.
-int StopForAxis(const StridedSliceParams &params, const ir::Shape &input_shape, int axis,
-                int start_for_axis)
-{
-  const auto end_mask = params.end_mask;
-  const auto shrink_axis_mask = params.shrink_axis_mask;
-  const auto *stop_indices = params.stop_indices;
-  const auto *strides = params.strides;
-
-  // Begin with the specified index
-  const bool shrink_axis = shrink_axis_mask & (1 << axis);
-  int stop = stop_indices[axis];
-
-  // When shrinking an axis, the end position does not matter (and can be
-  // incorrect when negative indexing is used, see Issue #19260). Always use
-  // start_for_axis + 1 to generate a length 1 slice, since start_for_axis has
-  // already been adjusted for negative indices.
-  if (shrink_axis)
-  {
-    stop = start_for_axis + 1;
-  }
-
-  // end_mask override
-  if (end_mask & (1 << axis))
-  {
-    if (strides[axis] > 0)
-    {
-      // Forward iteration - use the last element. These values will get
-      // clamped below
-      stop = std::numeric_limits<int>::max();
-    }
-    else
-    {
-      // Backward iteration - use the first element.
-      stop = std::numeric_limits<int>::lowest();
-    }
-  }
-
-  // Handle negative indices
-
-  const int axis_size = input_shape.dim(axis);
-  if (stop < 0)
-  {
-    stop += axis_size;
-  }
-
-  // Clamping
-  // Because the end index points one past the last element, we need slightly
-  // different clamping ranges depending on the direction.
-  if (strides[axis] > 0)
-  {
-    // Forward iteration
-    stop = Clamp(stop, 0, axis_size);
-  }
-  else
-  {
-    // Backward iteration
-    stop = Clamp(stop, -1, axis_size - 1);
-  }
-
-  return stop;
-}
-
-template <typename T>
-StridedSliceParams buildStridedSliceParams(const T *begin, const T *end, const T *strides,
-                                           const uint32_t begin_mask, const uint32_t end_mask,
-                                           const uint32_t shrink_axis_mask, const uint8_t rank)
-{
-  StridedSliceParams op_params;
-  op_params.start_indices_count = rank;
-  op_params.stop_indices_count = rank;
-  op_params.strides_count = rank;
-
-  for (int i = 0; i < op_params.strides_count; ++i)
-  {
-    op_params.start_indices[i] = begin[i];
-    op_params.stop_indices[i] = end[i];
-    op_params.strides[i] = strides[i];
-
-    assert(op_params.strides[i] != 0);
-  }
-
-  op_params.begin_mask = begin_mask;
-  op_params.ellipsis_mask = 0; // NYI
-  op_params.end_mask = end_mask;
-  op_params.new_axis_mask = 0; // NYI
-  op_params.shrink_axis_mask = shrink_axis_mask;
-
-  assert(sizeof(op_params.begin_mask) * 4 >= rank);
-
-  return op_params;
-}
-
-ir::Shape inferStridedSliceShape(const ir::Shape &input_shape, const StridedSliceParams &op_params,
-                                 uint32_t rank)
-{
-  ir::Shape out_shape;
-
-  for (uint32_t idx = 0; idx < rank; ++idx)
-  {
-    int32_t stride = op_params.strides[idx];
-    int32_t begin = StartForAxis(op_params, input_shape, idx);
-    int32_t end = StopForAxis(op_params, input_shape, idx, begin);
-
-    // When shrinking an axis, the end position does not matter (and can be
-    // incorrect when negative indexing is used, see Issue #19260). Always use
-    // begin + 1 to generate a length 1 slice, since begin has
-    // already been adjusted for negative indices by StartForAxis.
-    const bool shrink_axis = op_params.shrink_axis_mask & (1 << idx);
-    if (shrink_axis)
-    {
-      end = begin + 1;
-    }
-
-    int32_t dim_shape = std::ceil((end - begin) / static_cast<float>(stride));
-    dim_shape = dim_shape < 0 ? 0 : dim_shape;
-    if (!shrink_axis)
-    {
-      out_shape.append(dim_shape);
-    }
-  }
-
-  return out_shape;
-}
 
 void StaticInferer::visit(const ir::operation::StridedSlice &op)
 {
